board/lsi: Use the Right "ECC Busy" Bit in the NAND Driver
After initiating a NAND operation (write or erase) the NAND
controller registers must not be accessed until the ECC
calculations are complete.  To avoid doing this, the driver
reads a status pin on the controller (which is routed to a
general purpose register).  In the 3500 case, the driver was
reading the wrong bit.

Signed-off-by: John Jacques <john.jacques@lsi.com>