board/lsi: Axxia L3 Cache Option for Linux Boot
Linux modifies some instructions before switch from
the initial 1:1 mapping to the final mapping.  If the
L3 is enabled, this can cause some of the changes to
be lost.  This commit adds the option to leave the L3
in the SFONLY state and let Linux turn it on after
switch to the final address map.  To leave the L3 off,
define 'LEAVE_L3_OFF'.

Signed-off-by: John Jacques <john.jacques@lsi.com>