board/lsi: Axxia SSP Driver Changes
Axxia X9 and XLF contain a PL022 SSP controller, but
with a difference.  The Axxia SSP includes DMA support.

This commit makes a copy of drivers/spi/pl022_spi.c as
drivers/spi/axxia_pl022_spi.c and updates the Makefile
and header files to use it for X9 and XLF.  DMA support
will be added later.

Signed-off-by: John Jacques <john.jacques@intel.com>