Re-added possibility to power cycle the USB ports when switching on power to hubs.
This feature was removed from the u-boot mainline tree back in Nov. 2014.

Power cycling is recommended for xhci controller ports, but not required.
Real hardware works without doing power-cycle, because of internal workings of how writes to
registers are handled.

However, the USB model is modelled in a way that a power-cycle is not only recommended,
but required, since the default at reset in the model for the power line is high (power on).
State machine inside USB model is triggered by a transition from 0 to 1 on the power line, hence
we need to have a power-cycle for the simulated hardware.

This can be solved by either changing the u-boot to add power-cycling (this commit) or
change the default at reset in the model.

A define is implemented: CONFIG_AXXIA_USB_POWERCYCLE
Define this if you want powercycling of ports when power is switched on in hubs.
If not defined, software will only set the power lines for the ports high, assuming that ports
are already powered off.

Signed-off-by: George André <georgex.andre@intel.com>