board/axxia: USB Work-Around
To include the work-around, define CONFIG_AXXIA_USB_WA_54631.

-- First Issue --

The issue could be seen in Host application when the sequence below
occurs:

    1. The PHY Host performs an RX.Detect operation in the P3 state.
    2. The PHY TX Host output that is disabled is above common mode.

This issue is only seen in an extreme corner case of the process where
the PHY TX output when in P3 and not holding common mode floats
substantially above common mode. In P3 the TX output is not specified
or designed to hold a common mode voltage. The output voltage
typically naturally discharges to below common mode, if not the RX
detect pulse generated might not meet the specification and can cause
the RX.Detect function to incorrectly determine a receiver is not
present.

The sequence of the failure is as follows:

    1. The PHY is in P3.
    2. During the time the common mode is disabled, the TX output
       rises above common mode.
    3. An RX.Detect operation is initiated.
    4. The Detect pulse amplitude is not in the specification range.
    5. The RX.Detect result is that a receiver is not present.

Depending on the state of the LTSSM, the USB3.0 Phy will either fail
to detect a connected SS receiver and revert to HS mode or the link
partner will go to compliance.

Workaround is as follows:

Set controller bit GUSB3PIPECTRL[28] = 1 (Perform receiver detect in
P2) to allow the PHY to establish common mode before requesting a
receiver detect.

-- Second Issue --

The issue could be seen in host and device applications when the
sequence below occurs:

    1. Either of the link partners initiates a U1/U2 entry.
    2. An U1/U2 exit is initiated before the PHY has transition to U1/U2.

This issue is only seen when the link partner is in the U1/U2 state
and the noise during electrical idle on the link is decoded as legal
8b/10b coded data by the USB PHY receiver. This issue has only been
seen during yield analysis on a very small number of units in high
volume manufacturing of other devices using this IP. For the USB3 link
to successfully exit a low power state, it must first have entered a
low power state. The USB receiver relies on the occurrence of 8
successive 8b/10b decode errors to infer the link partner has entered
electrical idle. Under some occurrences 8 successive decode errors
have not been seen by the receiver before an exit has been initiated.

The sequence of the failure is as follows:

    1. Both link partners are present in U0.
    2. GUSB3PIPECTL[18] (DELAYP1TRANS) in the controller is set to “1”
       for Synopsys PHYs.
    3. Either of the link partners initiates an U1/U2 entry by
       transmitting the appropriate LGO.
    4. The receiving host/device responds with a LAU packet.
    5. The initiating link partner sends out an LPMA and transitions
       to electrical idle.
    6. The host/device controller (with the USB PHY) waits for 8
       successive decode errors, before requesting the PHY to transition
       from U0 to Ux. (GUSB3PIPECTL[18] =1).
    7. During electrical idle, in extreme corner cases, 8 successive
       decode errors are not measured by the PHY before an U1/U2 exit
       is initiated.
    8. When a local or remote exit is initiated, the host/device does
       not respond, since it has not yet entered the low power state.

If this issue occurs, the low power exit fails and the link
transitions to SS. Inactive or the PHY extends the idle time in the
lowpower state before exiting to Recovery thus reducing the data
throughput.

Workaround

1. In the controller set GUSB3PIPECTL[18] = 0. The controller, after
   receiving the LPMA will immediately request the PHY to transition
   from U0 to Ux power state and not wait for decode errors.
2. In the PHY set RX_SCOPE_LFPS_EN = 1 (Register 0x1026 value 0x0001)
   to enable a low pass filter in the LFPS detector circuit.

Signed-off-by: John Jacques <john.jacques@intel.com>