board/axxia: Allow Modification of the Counter Frequency Value
During performance testing, it is sometimes useful to see how much
affect OS overhead is having.  This is especially true when using the
emulation platform, as the clocks are quite a bit slower.  This commit
allows the cntfrq_el0 value to be multiplied on all cores by a factor
set in the environment, 'cntfrq_multiplier'.  Since sleeps/delays will
increase by this factor, system ticks will be delayed by that factor
as well.

Signed-off-by: John Jacques <john.jacques@intel.com>