board/axxia: Cleanup the Watchdog Code
Fix a calculation error when setting the watchdog
and print a warning if a longer that possible timeout
is requested (along with setting the maximum timeout).

Move all the common defines etc. to axxia.h.

Add functions to get and set the timeout in seconds.
After the PLLs/Clocks have been initialized, update
the watchdog timeout.

Signed-off-by: John Jacques <john.jacques@intel.com>