board/axxia: Fix Problems with Secure Boot
Incorrect offset was used to determine the encryption
state of the image.

When loading a signed image with secure boot disabled,
the stripped images was not written to the correct location.

Only check the encrypted flag if there is a SBB header.

Fail explicitly if secure boot is not enabled and the image is
encrypted.

Signed-off-by: John Jacques <john.jacques@intel.com>